fix attestation validation for electra (#8310)

+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/datastructures/operations/Attestation.java
@@ -43,7 +43,7 @@ public interface Attestation extends SszData, SszContainer {
 
   SszBitlist getAggregationBits();
 
-  int getFirstCommitteeIndex();
+  UInt64 getFirstCommitteeIndex();
 
   default Optional<SszBitvector> getCommitteeBits() {
     return Optional.empty();
